boilerplate.inc: 'v var cmd..' only removes a single LF
:	touch file1 $'file1\n' $'file1\n\n';
:	v tokill readlink -e $'file1\n\n';
:	rm "$tokill";

now removes $'file1\n\n' and not 'file1'.  Note that
this is one of the most dangerous traps in POSIX shells,
as following removes 'file1' and not $'file1\n\n':

:	rm "$(readlink -e $'file1\n\n')";

This even makes it very dangerous to call something like
basename or, even more dangerous, dirname on somerhing,
an attacker might have under control!